iio-ring-test: fix typo and clean-up
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@9649 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ user/blkfin-test/iio-ring-test/iio-ring.c
@@ -216,7 +216,7 @@ int main(int argc, char **argv)
 			for (k = 0; k < num_vals; k++) {
 				__s16 val = *(__s16 *)(&data[i*scan_size
 							     + (k)*2]);
-				printf("%04x ", val);
+				printf("%04x ", (unsigned short)val);
 			}
 			printf(" %lld\n",
 			       *(__s64 *)(&data[(i + 1)
